+++ store/primary/src/store/persist.rs
@@ -61,6 +61,16 @@ impl Persist {
         self.persist.discovery_settings(settings, span.into())
     }
 
+    /// Update the next_run of a cluster DiscoverySettings record.
+    ///
+    /// The new value is based on the current time + settings.interval.
+    pub fn next_discovery_run<S>(&self, settings: DiscoverySettings, span: S) -> Result<()>
+    where
+        S: Into<Option<SpanContext>>,
+    {
+        self.persist.next_discovery_run(settings, span.into())
+    }
+
     /// Creat or update a Node record.
     pub fn node<S>(&self, node: NodeModel, span: S) -> Result<()>
     where